+++ sync-for-payables/src/sdk/models/shared/createbillresponse.ts
@@ -134,7 +134,7 @@ export class CreateBillResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })