+++ sync-for-payables/src/sdk/models/shared/createjournalresponse.ts
@@ -138,7 +138,7 @@ export class CreateJournalResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })