+++ sync-for-payables/src/sdk/models/shared/createsupplierresponse.ts
@@ -120,7 +120,7 @@ export class CreateSupplierResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })